requirements.txt
# ============================================================
# Wall Street Bro - Python Dependencies
# Python 3.10+ required
# Install: pip install -r requirements.txt
# ============================================================

# --- Core / Config ---
python-dotenv>=1.0.1
pydantic>=2.10.1
pydantic-settings>=2.6.1

# --- LangGraph / LangChain Orchestration ---
langgraph>=0.2.25
langchain-core>=0.3.0
langchain-anthropic>=0.2.0

# --- AI SDKs ---
anthropic>=0.35.0
openai>=1.51.0          # xAI uses OpenAI-compatible client pointed at x.ai base URL

# --- Redis (async + sync) ---
redis>=5.0.8

# --- FastAPI Backend ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sse-starlette>=2.1.0    # Server-Sent Events support
websockets>=13.0

# --- Streamlit Frontend ---
streamlit>=1.39.0
streamlit-autorefresh>=1.0.1
plotly>=5.24.0
pandas>=2.2.3

# --- Alpaca Brokerage ---
alpaca-py>=0.31.0

# --- Database ---
sqlalchemy>=2.0.36
asyncpg>=0.29.0
alembic>=1.13.3

# --- MCP (Model Context Protocol) ---
mcp>=1.2.0

# --- HTTP / Networking ---
httpx>=0.27.2
aiohttp>=3.10.10

# --- Utilities ---
structlog>=24.4.0        # Structured logging
tenacity>=8.3.0,<9.0.0  # Retry logic (capped for langchain-core compatibility)

# --- Testing ---
pytest>=8.3.3
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
